libvterm from nix workaround
emacs-libvterm looks for libvterm installed on the system [1]. If there
is none, it downloads a vendored version [2] and calls `make` on it. The
`Makefile` in libvterm tries to use `glibtool` on darwin [3]. This works
fine when using the `libtool` from homebrew installed with a `g` prefix
[4], but it does not work when using `libtool` from nix without a
prefix.

If you wanted to use the vendored libvterm, you could build it manually
with the following. Inspired by [5,6].
```
cd ~/.emacs.d/.local/straight/build-28.2/vterm
mkdir build
cd build
cmake ..
make LIBTOOL=libtool
```

I tried to pass the `LIBTOOL=libtool` through the cmake used by
emacs-libvterm, but I could not find a way to pass it to the
BUILD_COMMAND [7].

In the end I figured I could install libvterm-neovim in nix, but the nix
cmake uses a hook [8,9] to change the paths of the libraries. This hook is
not applied to the cmake available in a normal shell, but seem to only
be used in derivations and nix-shell.

I guees you could fix it somewhere upstream or make a deriviation of,
but I'm in over my head. I just document a workaround 😮‍💨

[1] https://github.com/akermu/emacs-libvterm/blob/master/CMakeLists.txt#:~:text=option(-,USE_SYSTEM_LIBVTERM,-%22Use%20system%20libvterm
[2] https://github.com/akermu/emacs-libvterm/blob/master/CMakeLists.txt#:~:text=endif()-,ExternalProject_add,-(libvterm
[3] https://bazaar.launchpad.net/~libvterm/libvterm/trunk/view/head:/Makefile#L1
[4] https://github.com/Homebrew/homebrew-core/blob/master/Formula/libtool.rb#:~:text=args%20%3C%3C%20%22-,%2D%2Dprogram%2Dprefix%3Dg,-%22%20if%20OS
[5] https://github.com/akermu/emacs-libvterm/issues/169#issuecomment-568668946
[6] https://github.com/akermu/emacs-libvterm/issues/169#issuecomment-846248073
[7] https://github.com/akermu/emacs-libvterm/blob/master/CMakeLists.txt#:~:text=BUILD_COMMAND%20%24%7BLIBVTERM_BUILD_COMMAND%7D
[8] https://nixos.wiki/wiki/C#cmake
[9] https://github.com/NixOS/nixpkgs/blob/master/pkgs/development/tools/build-managers/cmake/setup-hook.sh