promt is a function, not a command
In d07efaa649d90673fca4595a00f4ac46983cf5cd I fixed an issue with
missing newlines. In 92527551e3a5b64bffae4f543f4870d25a0e963e I tried to
amended that so it did not fail on non-nix systems, but I just disabled
the fix as `prompt` is a `function`, not a `command`. This should fix
it.