Use nixpkg supplied emacs-macport
I have looked at multiple different alternatives and I haved changed
solutions multiple times. I will not try to keep the various
installation methods and bugs on
https://docs.google.com/spreadsheets/d/1ZTwuL0W0lKMCu_YrAUiUcCiV9shdHrbszjdDjoKJ918/edit?gid=0#gid=0

This removes:

- the homebrew railwaycat/emacmacsport that does not support native comp
  and only have a 1000 file descriptors
- macports emacs-mac-app that works, but requires macports as a
  dependency and take multiple hours to compile. The emacsclient from
  nix did not work.

I think the reason the nixpkg emacs and emacs-macport was disqualified
earlier was because of missing .app bundles and Spotlight search. With
mac-app-util, this not a problem now.

There are two outstanding issues:
- I have not tested the number file descriptors, but if it is a problem,
  I think I can patch it.
- /run/current-system/sw/bin/emacs creates an emacs window that does not
  aquire focus. macOS send all keypresses to the iTerm window that
  spawned it.