Use emacs-mac 29.4
This is an attempt to fix a fleeting bug where emacs somtimes freezes
with 100% CPU usage. A bug that sounds like the same is described in
https://www.reddit.com/r/emacs/comments/1ba3819/comment/ku49lmk/.

I have to use this for some time to see if the comes up again.

All the emacs packaging bugs are tracked on https://docs.google.com/spreadsheets/d/1ZTwuL0W0lKMCu_YrAUiUcCiV9shdHrbszjdDjoKJ918/edit?gid=0#gid=0